+++ packages/pak-files-parser/src/main.ts
@@ -9,6 +9,7 @@ import { NpcPortraitsImageProcessor } from "./app/image-processors/npc-portraits
 import { StringTable } from "./util/string-table.class";
 import {
     AnimalShopData,
+    AvailableLanguage,
     AvailableLanguages,
     CookingRecipe,
     CraftingRecipe,
@@ -68,7 +69,7 @@ if (environment.isBeta) {
 }
 NPCDbGenerator.AdditionalNpcAppearances = additionalNPCOutfitsMappings;
 
-(AvailableLanguages).forEach((lang, langIndex) => {
+(['en'] as AvailableLanguage[]).forEach((lang, langIndex) => {
     Logger.info(`Generators for "${lang}" starting...`);
     StringTable.defaultLang = lang;
 
@@ -473,10 +474,12 @@ NPCDbGenerator.AdditionalNpcAppearances = additionalNPCOutfitsMappings;
         })
 
 
-        if (langIndex === AvailableLanguages.length - 1) {
+        if (true || langIndex === AvailableLanguages.length - 1) {
             Logger.info('Create dashboard files')
-            DashboardFilesCreation(dbItems);
+            DashboardFilesCreation(dbItems, npcDbMap, calendarDbMap, generatorResults['gift-preferences']);
             Logger.info('Creating dashboard files done')
+
+            process.exit()
         }
 
     } catch (e) {